# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Rakshak AI - Crime Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /firs:
    get:
      operationId: listFirs
      tags: [firs]
      summary: List all FIRs with filters
      parameters:
        - name: crimeType
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: FIR list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FIR"
    post:
      operationId: createFir
      tags: [firs]
      summary: Create a new FIR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FIRInput"
      responses:
        "201":
          description: Created FIR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FIR"
  /firs/{id}:
    get:
      operationId: getFir
      tags: [firs]
      summary: Get a specific FIR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: FIR details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FIR"
        "404":
          description: Not found
  /firs/{id}/summary:
    get:
      operationId: getFirSummary
      tags: [firs]
      summary: AI-generated summary and investigation timeline for an FIR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: FIR AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirSummary"
  /firs/{id}/similar:
    get:
      operationId: getSimilarFirs
      tags: [firs]
      summary: Find similar past FIR cases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar FIRs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimilarCase"

  # ─── Accused / Offender Profiles ─────────────────────────────────────────
...
components:
  schemas:
...
    FIR:
      type: object
      properties:
        id:
          type: integer
        firNumber:
          type: string
        crimeType:
          type: string
        district:
          type: string
        policeStation:
          type: string
        dateOfOccurrence:
          type: string
        location:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        description:
          type: string
        status:
          type: string
          enum: [open, under_investigation, chargesheet_filed, closed]
        victimCount:
          type: integer
        accusedCount:
          type: integer
        officerInCharge:
          type: string
        createdAt:
          type: string
      required:
        - id
        - firNumber
        - crimeType
        - district
        - policeStation
        - dateOfOccurrence
        - location
        - description
        - status
        - victimCount
        - accusedCount
        - officerInCharge
        - createdAt
    FIRInput:
      type: object
      properties:
        firNumber:
          type: string
        crimeType:
          type: string
        district:
          type: string
        policeStation:
          type: string
        dateOfOccurrence:
          type: string
        location:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        description:
          type: string
        officerInCharge:
          type: string
      required:
        - firNumber
        - crimeType
        - district
        - policeStation
        - dateOfOccurrence
        - location
        - description
        - officerInCharge
    FirSummary:
      type: object
      properties:
        firId:
          type: integer
        aiSummary:
          type: string
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        recommendations:
          type: array
          items:
            type: string
        keyFindings:
          type: array
          items:
            type: string
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
      required:
        - firId
        - aiSummary
        - timeline
        - recommendations
        - keyFindings
        - riskLevel
    TimelineEvent:
      type: object
      properties:
        date:
          type: string
        event:
          type: string
        type:
          type: string
      required:
        - date
        - event
        - type
    SimilarCase:
      type: object
      properties:
        firId:
          type: integer
        firNumber:
          type: string
        crimeType:
          type: string
        similarity:
          type: number
        matchReasons:
          type: array
          items:
            type: string
        outcome:
          type: string
      required:
        - firId
        - firNumber
        - crimeType
        - similarity
        - matchReasons
        - outcome

    # Accused
...